v1.2
较为稳定的pcm音频声纹注册，验证版本
五个主要接口：
1）用户注册：需客户端上传必要的用户信息，f返回json包括statues_code=1 文件上传失败   ,2  ivector计算错误
,4 用户已注册或者person_id无效，或person_id为null  ,8  提交表单缺少必要信息
2）用户验证
3）用户注销
4）用户查询
5）用户更新

+++ src/com/pingan/servlet/ValidationServlet.java
@@ -158,7 +158,7 @@ public class ValidationServlet extends HttpServlet {
 
 			System.out.println("validation_result=" + score);
 
-			int threshold = 0;
+			float threshold = 0;
 			switch (Constant.SCORE_MODE) {
 			case PLDA:
 				threshold = Constant.PLDA_THRESHOLD;
